added controllers for routes, added seeding json, cleaned up server file
